factor out lexer
