+++ src/model.rs
@@ -1,3 +1,4 @@
+use super::lexer::Lexer;
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 use std::path::{Path, PathBuf};
@@ -234,65 +235,4 @@ fn compute_idf(t: &str, n: usize, df: &DocFreq) -> f32 {
     let nume = n as f32;
     let deno = df.get(t).cloned().unwrap_or(1) as f32;
     (nume / deno).log10()
-}
-
-pub struct Lexer<'a> {
-    content: &'a [char],
-}
-
-impl<'a> Lexer<'a> {
-    pub fn new(content: &'a [char]) -> Self {
-        Self { content }
-    }
-
-    fn trim_left(&mut self) {
-        while !self.content.is_empty() && self.content[0].is_whitespace() {
-            self.content = &self.content[1..];
-        }
-    }
-
-    fn chop(&mut self, n: usize) -> &'a [char] {
-        let token = &self.content[0..n];
-        self.content = &self.content[n..];
-        token
-    }
-
-    fn chop_while<P>(&mut self, mut predicate: P) -> &'a [char]
-    where
-        P: FnMut(&char) -> bool,
-    {
-        let mut n = 0;
-        while n < self.content.len() && predicate(&self.content[n]) {
-            n += 1;
-        }
-        self.chop(n)
-    }
-
-    pub fn next_token(&mut self) -> Option<String> {
-        self.trim_left();
-        if self.content.len() == 0 {
-            return None;
-        }
-
-        if self.content[0].is_numeric() {
-            return Some(self.chop_while(|x| x.is_numeric()).iter().collect());
-        }
-
-        if self.content[0].is_alphabetic() {
-            return Some(
-                self.chop_while(|x| x.is_alphanumeric())
-                    .iter()
-                    .map(|x| x.to_ascii_uppercase())
-                    .collect(),
-            );
-        }
-        return Some(self.chop(1).iter().collect());
-    }
-}
-
-impl<'a> Iterator for Lexer<'a> {
-    type Item = String;
-    fn next(&mut self) -> Option<Self::Item> {
-        self.next_token()
-    }
 }
